+++ docs/scripts/write_bitmask_tables.py
@@ -31,9 +31,9 @@ if __name__ == '__main__':
 
     # Tables to write:
     #
-    from mangadap.drpfits import DRPQuality3DBitMask
+    from mangadap.util.drpfits import DRPQuality3DBitMask
     write_bitmask(DRPQuality3DBitMask, path)
-    from mangadap.drpfits import DRPFitsBitMask
+    from mangadap.util.drpfits import DRPFitsBitMask
     write_bitmask(DRPFitsBitMask, path)
 
     from mangadap.dapfits import DAPQualityBitMask