+++ mangadap/datacube/manga.py
@@ -15,8 +15,6 @@ import time
 import os
 import warnings
 
-from configparser import ConfigParser
-
 from IPython import embed
 
 import numpy
@@ -25,7 +23,7 @@ from astropy.io import fits
 from astropy.wcs import WCS
 
 from ..config import defaults
-from ..drpfits import DRPFitsBitMask
+from ..util.drpfits import DRPFits, DRPFitsBitMask
 from ..util.parser import DefaultConfig
 from ..util.constants import DAPConstants
 from ..util.covariance import Covariance
@@ -33,15 +31,16 @@ from ..util.filter import interpolate_masked_vector
 from ..spectra import MaNGARSS
 from .datacube import DataCube
 
-class MaNGADataCube(DataCube):
+class MaNGADataCube(DRPFits, DataCube):
     r"""
     Container class for a MaNGA datacube.
 
-    For additional description and attributes, see
-    :class:`mangadap.datacube.datacube.DataCube`.
+    For additional description and attributes, see the two base
+    classes.
 
     See :func:`from_plateifu` to instantiate using the plate and ifu
-    numbers.
+    numbers. See :func:`from_config` to instantiate from a
+    configuration file.
 
     Note that ``z``, ``vdisp``, ``ell``, ``pa``, and ``reff`` are
     saved to the :attr:`meta` dictionary.
@@ -91,11 +90,11 @@ class MaNGADataCube(DataCube):
             raise FileNotFoundError('File does not exist: {0}'.format(ifile))
 
         # Parse the relevant information from the filename
-        self.directory_path, self.file = os.path.split(os.path.abspath(ifile)) 
-        self.plate, self.ifudesign, log = self.file.split('-')[1:4]
-        self.plate = int(self.plate)
-        self.ifudesign = int(self.ifudesign)
-        log = 'LOG' in log
+        directory_path = os.path.split(os.path.abspath(ifile))[0]
+        plate, ifudesign, log = ifile.split('-')[1:4]
+        # Instantiate the DRPFits base
+        DRPFits.__init__(self, int(plate), int(ifudesign), 'CUBE', log='LOG' in log,
+                         directory_path=directory_path)
 
         # Collect the metadata into a dictionary
         meta = {}
@@ -113,7 +112,7 @@ class MaNGADataCube(DataCube):
                           ' between masked (values greater than 0) and unmasked (values of 0).')
             bitmask = None
 
-        # Open the file and initialize the base class
+        # Open the file and initialize the DataCube base class
         with fits.open(ifile) as hdu:
             # Read covariance first
             covar = None if covar_ext is None \
@@ -126,8 +125,7 @@ class MaNGADataCube(DataCube):
             # NOTE: Need to keep a log of what spectral resolution
             # vectors were used so that the same vectors are read from
             # the RSS file, if/when it is loaded.
-            self.sres_ext, sres = MaNGADataCube.spectral_resolution(hdu, ext=sres_ext,
-                                                                    fill=sres_fill)
+            self.sres_ext, sres = DRPFits.spectral_resolution(hdu, ext=sres_ext, fill=sres_fill)
             self.sres_fill = sres_fill
             sres = sres.filled(0.0)
 
@@ -140,11 +138,11 @@ class MaNGADataCube(DataCube):
             # is expected to be okay because numpy array flattening
             # always performs a C-like flattening, even if the memory
             # storage is Fortran contiguous.
-            super(MaNGADataCube, self).__init__(hdu['FLUX'].data.T, wave=hdu['WAVE'].data,
-                                                ivar=hdu['IVAR'].data.T, mask=hdu['MASK'].data.T,
-                                                bitmask=bitmask, sres=sres.T, covar=covar,
-                                                wcs=WCS(header=fluxhdr, fix=True), pixelscale=0.5,
-                                                log=log, meta=meta, prihdr=prihdr, fluxhdr=fluxhdr)
+            DataCube.__init__(self, hdu['FLUX'].data.T, wave=hdu['WAVE'].data,
+                              ivar=hdu['IVAR'].data.T, mask=hdu['MASK'].data.T, bitmask=bitmask,
+                              sres=sres.T, covar=covar, wcs=WCS(header=fluxhdr, fix=True),
+                              pixelscale=0.5, log=self.samp == 'LOG', meta=meta, prihdr=prihdr,
+                              fluxhdr=fluxhdr)
         print('Reading MaNGA datacube data ... DONE')
 
         # Try to use the header to set the DRP version
@@ -159,122 +157,83 @@ class MaNGADataCube(DataCube):
             warnings.warn('Default reduction path does not match file path.  May not be able to '
                           'find paired RSS file if requested.')
 
+    # TODO: Include a class method that instantiates from (or wraps a Marvin Cube)
     @staticmethod
-    def spectral_resolution_extension(hdu, ext=None):
+    def build_file_name(plate, ifudesign, log=True):
         """
-        Determine the spectral resolution channel to use.
+        Return the name of the DRP datacube file.
 
-        Precedence follows this order: ``PREDISP``, ``PRESPECRES``,
-        ``DISP``, ``SPECRES``.
+        This is a simple wrapper for
+        :func:`mangadap.util.drpfits.DRPFits.build_file_name`,
+        specific to the datacube.
 
         Args:
-            hdu (`astropy.io.fits.HDUList`):
-                The opened MaNGA DRP file.
-            ext (:obj:`str`, optional):
-                Specify the extension with the spectral estimate to
-                use. Should be in None, ``PREDISP``, ``PRESPECRES``,
-                ``DISP``, or ``SPECRES``. The default is None, which
-                means it will return the extension found first in the
-                order above. None is returned if none of the
-                extensions are present.
+            plate (:obj:`int`):
+                Plate number
+            ifudesign (:obj:`int`):
+                IFU design
+            log (:obj:`bool`, optional):
+                Use the spectra that are logarithmically sampled in
+                wavelength. If False, sampling is linear in
+                wavelength.
 
         Returns:
-            :obj:`str`: The name of the preferred extension to use.
+            :obj:`str`: The relevant file name.
         """
-        available = [h.name for h in hdu if h.name in ['PREDISP', 'DISP', 'PRESPECRES', 'SPECRES']]
-        _ext = ext
-        if ext is None:
-            _ext = 'PREDISP'
-            if _ext not in available:
-                _ext = 'PRESPECRES'
-            if _ext not in available:
-                _ext = 'DISP'
-            if _ext not in available:
-                _ext = 'SPECRES'
-        return None if _ext not in available else _ext
+        return DRPFits.build_file_name(plate, ifudesign, 'CUBE', log=log)
 
     @staticmethod
-    def spectral_resolution(hdu, ext=None, fill=False, median=False):
+    def default_paths(plate, ifudesign, log=True, drpver=None, redux_path=None,
+                      directory_path=None):
         """
-        Return the spectral resolution at each spatial and spectral
-        position.
+        Construct the default path and file name with the MaNGA
+        datacube.
 
-        See :func:`spectral_resolution_extension` for a description
-        of the precedence used when ``ext`` is None.
+        This is a simple wrapper for
+        :func:`mangadap.util.drpfits.DRPFits.default_paths`, specific
+        to the datacube files.
 
         Args:
-            hdu (`astropy.io.fits.HDUList`):
-                The opened MaNGA DRP file.
-            ext (:obj:`str`, optional):
-                Specify the extension with the spectral estimate to
-                use. See :func:`spectral_resolution_extension`.
-            fill (:obj:`bool`, optional):
-                Fill masked values by interpolation.  Default is to
-                leave masked pixels in returned array.
-            median (:obj:`bool`, optional):
-                Return a single vector with the median spectral
-                resolution instead of a per spectrum array.  When using
-                the `SPECRES` extension, this just returns the vector
-                provided by the DRP file; when using the `DISP`
-                extension, this performs a masked median across the
-                array and then interpolates any wavelengths that were
-                masked in all vectors.
+            plate (:obj:`int`):
+                Plate number
+            ifudesign (:obj:`int`):
+                IFU design
+            log (:obj:`bool`, optional):
+                Use the spectra that are logarithmically sampled in
+                wavelength. If False, sampling is linear in
+                wavelength.
+            drpver (:obj:`str`, optional):
+                DRP version, which is used to define the default DRP
+                redux path. Default is defined by
+                :func:`mangadap.config.defaults.drp_version`
+            redux_path (:obj:`str`, optional):
+                The path to the top level directory containing the
+                DRP output files for a given DRP version. Default is
+                defined by
+                :func:`mangadap.config.defaults.drp_redux_path`.
+            directory_path (:obj:`str`, optional):
+                The exact path to the DRP file. Default is defined by
+                :func:`mangadap.config.defaults.drp_directory_path`.
+                Providing this ignores anything provided for
+                ``drpver`` or ``redux_path``.
 
         Returns:
-            :obj:`tuple`: Returns a :obj:`str` with the name of the
-            extension used for the spectral resolution measurements
-            and a `numpy.ma.MaskedArray`_ with the spectral
-            resolution data. Even if interpolated such that there
-            should be no masked values, the function returns a masked
-            array. Array contains the spectral resolution (:math:`R =
-            \lambda/\Delta\lambda`) pulled from the DRP file.
+            :obj:`tuple`: Two strings with the default path to and
+            name of the DRP data file.
         """
-        # Determine which spectral resolution element to use
-        _ext = MaNGADataCube.spectral_resolution_extension(hdu, ext=ext)
-        # If no valid extension, raise an exception
-        if ext is None and _ext is None:
-            raise ValueError('No valid spectral resolution extension.')
-        if ext is not None and _ext is None:
-            raise ValueError('No extension: {0}'.format(ext))
-            
-        # Build the spectral resolution vectors
-        if 'SPECRES' in _ext:
-            sres = numpy.ma.MaskedArray(hdu[_ext].data, mask=numpy.invert(hdu[_ext].data > 0))
-            if fill:
-                sres = numpy.ma.MaskedArray(interpolate_masked_vector(sres))
-            if not median:
-                spatial_shape = hdu['FLUX'].data.shape[1:][::-1]
-                sres = numpy.ma.tile(sres, spatial_shape + (1,)).T
-            return _ext, sres
-
-        # Otherwise dealing with the DISP cube
-        sres = numpy.ma.MaskedArray(hdu[_ext].data)
-        # Mask any non-positive value
-        sres[numpy.invert(sres > 0)] = numpy.ma.masked
-        # Convert from sigma in angstroms to spectral resolution
-        # (based on FWHM)
-        sres = numpy.ma.divide(hdu['WAVE'].data[:,None,None], sres) / DAPConstants.sig2fwhm
-        # Interpolate over any masked values
-        if fill:
-            outshape = sres.shape
-            sres = numpy.ma.MaskedArray(
-                        numpy.ma.apply_along_axis(interpolate_masked_vector, 0,
-                                                  sres.reshape(outshape[0], -1))).reshape(outshape)
-        if median:
-            sres = numpy.ma.median(sres.reshape(outshape[0],-1), axis=1)
-        return _ext, sres
+        return DRPFits.default_paths(plate, ifudesign, 'CUBE', log=log, drpver=drpver,
+                                     redux_path=redux_path, directory_path=directory_path)
 
     @classmethod
     def from_plateifu(cls, plate, ifudesign, log=True, drpver=None, redux_path=None,
                       directory_path=None, **kwargs):
         """
-        Construct a MaNGA datacube object based on its plate-ifu
-        designation.
+        Construct a :class:`mangadap.datacube.manga.MaNGADataCube`
+        object based on its plate-ifu designation.
 
-        The provided plate, ifudesign, and boolean for the log
-        binning are used to construct the name of the file. The DRP
-        version and reduction and directory paths are used to
-        construct the directory with the file.
+        This is a simple wrapper function that calls
+        :func:`default_paths` to construct the file names, and then
+        calls the class instantiation method.
 
         Args:
             plate (:obj:`int`):
@@ -300,180 +259,13 @@ class MaNGADataCube(DataCube):
                 ``drpver`` or ``redux_path``.
             **kwargs:
                 Keyword arguments passed directly to the primary
-                instantiation method; see :class:`MaNGADataCube`.
+                instantiation method; see
+                :class:`mangadap.datacube.manga.MaNGADataCube` or
+                :class:`mangadap.spectra.manga.MaNGARSS`.
         """
-        # Set the attributes, forcing a known type
-        _plate = int(plate)
-        _ifudesign = int(ifudesign)
-
-        # Setup the directory path.
-        if directory_path is None:
-            directory_path = defaults.drp_directory_path(_plate, drpver=drpver,
-                                                         redux_path=redux_path)
-        return cls(os.path.join(directory_path,
-                                MaNGADataCube.build_file_name(_plate, _ifudesign, log=log)),
-                   **kwargs)
-
-    @classmethod
-    def from_config(cls, cfgfile, drpver=None, redux_path=None, directory_path=None):
-        """
-        Construct a MaNGA datacube object using a configuration file.
-
-        The format of the configuration file is:
-
-        .. todo::
-
-            Fill this in.
-
-        Args:
-            cfgfile (:obj:`str`):
-                Configuration file
-            drpver (:obj:`str`, optional):
-                DRP version, which is used to define the default DRP
-                redux path. Default is defined by
-                :func:`mangadap.config.defaults.drp_version`.
-                Overrides any value in the configuration file.
-            redux_path (:obj:`str`, optional):
-                The path to the top level directory containing the
-                DRP output files for a given DRP version. Default is
-                defined by
-                :func:`mangadap.config.defaults.drp_redux_path`.
-                Overrides any value in the configuration file.
-            directory_path (:obj:`str`, optional):
-                The exact path to the DRP file. Default is defined by
-                :func:`mangadap.config.defaults.drp_directory_path`.
-                Providing this ignores anything provided for
-                ``drpver`` or ``redux_path``. Overrides any value i
-                the configuration file.
-        """
-        # Read the configuration file
-        cfg = DefaultConfig(cfgfile, interpolate=True)
-
-        # Set the attributes, forcing a known type
-        plate = cfg.getint('plate')
-        ifu = cfg.getint('ifu')
-        if plate is None or ifu is None:
-            raise ValueError('Configuration file must define the plate and IFU.')
-        log = cfg.getbool('log', default=True)
-
-        # Overwrite what's in the file with the method keyword arguments
-        _drpver = cfg.get('drpver') if drpver is None else drpver
-        _redux_path = cfg.get('redux_path') if redux_path is None else redux_path
-        _directory_path = cfg.get('directory_path') if directory_path is None else directory_path
-
-        # Get the other possible keywords
-        # TODO: Come up with a better way to do this
-        kwargs = {}
-        kwargs['sres_ext'] = cfg.get('sres_ext')
-        kwargs['sres_fill'] = cfg.getbool('sres_fill', default=True)
-        kwargs['covar_ext'] = cfg.get('covar_ext')
-        kwargs['z'] = cfg.getfloat('z')
-        kwargs['vdisp'] = cfg.getfloat('vdisp')
-        kwargs['ell'] = cfg.getfloat('ell')
-        kwargs['pa'] = cfg.getfloat('pa')
-        kwargs['reff'] = cfg.getfloat('reff')
-
-        return cls.from_plateifu(plate, ifu, log=log, drpver=_drpver, redux_path=_redux_path,
-                                 directory_path=_directory_path, **kwargs)
-
-    @staticmethod
-    def write_config(ofile, plate, ifudesign, log=True, z=None, vdisp=None, ell=None, pa=None,
-                     reff=None, sres_ext=None, sres_fill=None, covar_ext=None, drpver=None,
-                     redux_path=None, directory_path=None, overwrite=True):
-        """
-        Write the configuration file that can be used to instantiate
-        the datacube.
-
-        See :func:`from_config`.
-
-        Args:
-            ofile (:obj:`str`, optional):
-                Name of the configuration file.
-            plate (:obj:`int`):
-                Plate number
-            ifudesign (:obj:`int`):
-                IFU design
-            log (:obj:`bool`, optional):
-                Use the datacube that is logarithmically binned in
-                wavelength.
-            z (:obj:`float`, optional):
-                Estimated bulk redshift. If None, some of the DAP
-                analysis modules will fault.
-            vdisp (:obj:`float`, optional):
-                Estimated velocity dispersion. If None, some of the
-                DAP analysis modules will assume an initial guess of
-                100 km/s.
-            ell (:obj:`float`, optional):
-                Characteristic isophotal ellipticity (1-b/a). If
-                None, some of the DAP modules will issue a warning
-                and continue by assuming ``ell=0``.
-            pa (:obj:`float`, optional):
-                Characteristic isophotal position angle (through E
-                from N). If None, some of the DAP modules will issue
-                a warning and continue by assuming ``pa=0``.
-            reff (:obj:`float`, optional):
-                Effective (half-light) radius in arcsec. If None,
-                some of the DAP modules will issue a warning and
-                continue by assuming ``reff=1``.
-            sres_ext (:obj:`str`, optional):
-                The extension to use when constructing the spectral
-                resolution vectors. See :func:`spectral_resolution`.
-            sres_fill (:obj:`bool`, optional):
-                Fill masked values by interpolation. Default is to
-                leave masked pixels in returned array.
-            covar_ext (:obj:`str`, optional):
-                Extension to use as the single spatial correlation
-                matrix for all wavelength channels, read from the DRP
-                file. For generating the covariance matrix directly
-                for an arbitrary wavelength channel using the RSS
-                file, see
-                :func:`mangadap.datacube.datacube.DataCube.covariance_matrix`.
-            drpver (:obj:`str`, optional):
-                DRP version, which is used to define the default DRP
-                redux path. Default is defined by
-                :func:`mangadap.config.defaults.drp_version`
-            redux_path (:obj:`str`, optional):
-                The path to the top level directory containing the
-                DRP output files for a given DRP version. Default is
-                defined by
-                :func:`mangadap.config.defaults.drp_redux_path`.
-            directory_path (:obj:`str`, optional):
-                The exact path to the DRP file. Default is defined by
-                :func:`mangadap.config.defaults.drp_directory_path`.
-                Providing this ignores anything provided for
-                ``drpver`` or ``redux_path``.
-            overwrite (:obj:`bool`, optional):
-                Overwrite any existing parameter file.
-        """
-        if os.path.exists(ofile) and not overwrite:
-            raise FileExistsError('Configuration file already exists; to overwrite, set '
-                                  'overwrite=True.')
-
-        # Build the configuration data
-        cfg = ConfigParser(allow_no_value=True)
-        cfg['default'] = {'drpver': drpver,
-                          'redux_path': redux_path,
-                          'directory_path': directory_path,
-                          'plate': str(plate),
-                          'ifu': str(ifudesign),
-                          'log': str(log),
-                          'sres_ext': sres_ext,
-                          'sres_fill': sres_fill,
-                          'covar_ext': covar_ext,
-                          'z': None if z is None else '{0:.7e}'.format(z),
-                          'vdisp': None if vdisp is None else '{0:.7e}'.format(vdisp),
-                          'ell': None if ell is None else '{0:.7e}'.format(ell),
-                          'pa': None if pa is None else '{0:.7e}'.format(pa),
-                          'reff': None if reff is None else '{0:.7e}'.format(reff)}
-
-        # Write the configuration file
-        with open(ofile, 'w') as f:
-            f.write('# Auto-generated configuration file\n')
-            f.write('# {0}\n'.format(time.strftime("%a %d %b %Y %H:%M:%S",time.localtime())))
-            f.write('\n')
-            cfg.write(f)
-
-    # TODO: Include a class method that instantiates from (or wraps a Marvin Cube)
+        path, file_name = cls.default_paths(int(plate), int(ifudesign), log=log, drpver=drpver,
+                                            redux_path=redux_path, directory_path=directory_path)
+        return cls(os.path.join(directory_path, file_name), **kwargs)
 
     def load_rss(self, force=False):
         """
@@ -501,7 +293,7 @@ class MaNGADataCube(DataCube):
         if self.rss is not None and not force:
             return
 
-        rss_file = MaNGARSS.build_file_name(self.plate, self.ifudesign, log=self.log)
+        rss_file = DRPFits.build_file_name(self.plate, self.ifudesign, 'RSS', log=self.log)
         rss_file_path = os.path.join(self.directory_path, rss_file)
         if not os.path.isfile(rss_file_path):
             # Not in this directory.  Check the nominal directory
@@ -514,89 +306,6 @@ class MaNGADataCube(DataCube):
             raise FileNotFoundError('Could not find RSS file.')
 
         self.rss = MaNGARSS(rss_file_path, sres_ext=self.sres_ext, sres_fill=self.sres_fill)
-
-    @staticmethod
-    def build_file_name(plate, ifudesign, log=True):
-        """
-        Return the name of the DRP datacube file.
-
-        Args:
-            plate (:obj:`int`):
-                Plate number
-            ifudesign (:obj:`int`):
-                IFU design
-            log (:obj:`bool`, optional):
-                Use the datacube that is logarithmically binned in
-                wavelength.
-
-        Returns:
-            :obj:`str`: The relevant file name.
-        """
-        mode = '{0}CUBE'.format('LOG' if log else 'LIN')
-        return '{0}.fits.gz'.format(defaults.manga_fits_root(plate, ifudesign, mode))
-
-    def file_path(self):
-        """Return the full path to the DRP datacube file."""
-        return os.path.join(self.directory_path, self.file)
-
-    @staticmethod
-    def do_not_use_flags():
-        """Return the maskbit names that should not be used."""
-        return ['DONOTUSE', 'FORESTAR']
-
-    @staticmethod
-    def do_not_fit_flags():
-        """Return the maskbit names that should not be fit."""
-        return ['DONOTUSE', 'FORESTAR']
-
-    @staticmethod
-    def do_not_stack_flags():
-        """Return the maskbit names that should not be stacked."""
-        return ['DONOTUSE', 'FORESTAR']
-
-    @staticmethod
-    def default_paths(plate, ifudesign, log=True, drpver=None, redux_path=None,
-                      directory_path=None, output_file=None):
-        """
-        Return the primary directory and file name with the DRP fits
-        LOG-binned file.
-
-        Args:
-            plate (:obj:`int`):
-                Plate number
-            ifudesign (:obj:`int`):
-                IFU design
-            log (:obj:`bool`, optional):
-                Use the datacube that is logarithmically binned in
-                wavelength.
-            drpver (:obj:`str`, optional):
-                DRP version, which is used to define the default DRP
-                redux path. Default is defined by
-                :func:`mangadap.config.defaults.drp_version`
-            redux_path (:obj:`str`, optional):
-                The path to the top level directory containing the
-                DRP output files for a given DRP version. Default is
-                defined by
-                :func:`mangadap.config.defaults.drp_redux_path`.
-            directory_path (:obj:`str`, optional):
-                The exact path to the DRP file. Default is defined by
-                :func:`mangadap.config.defaults.drp_directory_path`.
-                Providing this ignores anything provided for
-                ``drpver`` or ``redux_path``.
-            output_file (:obj:`str`, optional):
-                The name of the file with the DRP data. Default set
-                by :func:`MaNGADataCube.build_file_name`.
-
-        Returns:
-            :obj:`str`: Two strings with the path to and name of the DRP
-            data file.
-        """
-        _directory_path = defaults.drp_directory_path(plate, drpver=drpver,
-                                                      redux_path=redux_path) \
-                                if directory_path is None else directory_path
-        _output_file = MaNGADataCube.build_file_name(plate, ifudesign, log=log) \
-                            if output_file is None else output_file
-        return _directory_path, _output_file
 
     def mean_sky_coordinates(self, center_coo=None, offset='OBJ'):
         """